Azure-K8S-On-Arc/python-api/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.95.0
uvicorn[standard]==0.20.0
gunicorn==20.1.0
redis==4.5.1
pydantic==1.10.5
python-dotenv==1.0.0